+++ frontend/src/components/user/SectionInfoPreview/index.tsx
@@ -1,5 +1,6 @@
-import { useEffect, useRef, useState } from 'react';
 import { FaMapMarkedAlt } from 'react-icons/fa';
+
+import useLazyImage from '@/hooks/useLazyImage';
 
 import styles from './styles';
 
@@ -7,31 +8,6 @@ interface SectionInfoPreviewProps {
   imageUrl: string;
   onClick: () => void;
 }
-
-const useLazyImage = () => {
-  const lazyImageRef = useRef<HTMLImageElement>(null);
-  const observerRef = useRef<IntersectionObserver>();
-  const [isLoaded, setIsLoaded] = useState(false);
-
-  const intersectionCallBack = (entries: IntersectionObserverEntry[], io: IntersectionObserver) => {
-    entries.forEach(entry => {
-      if (entry.isIntersecting) {
-        io.unobserve(entry.target);
-        setIsLoaded(true);
-      }
-    });
-  };
-
-  useEffect(() => {
-    if (!observerRef.current) {
-      observerRef.current = new IntersectionObserver(intersectionCallBack);
-    }
-
-    lazyImageRef.current && observerRef.current.observe(lazyImageRef.current);
-  }, []);
-
-  return { isLoaded, lazyImageRef };
-};
 
 const SectionInfoPreview: React.FC<SectionInfoPreviewProps> = ({ imageUrl, onClick }) => {
   const { isLoaded, lazyImageRef } = useLazyImage();
